Added Outbreak Service. Added `react-native-fs`. Decode the outbreak payload downloaded from server. Check outbreakEvents against checkInHistory.
