+++ package.json
@@ -12,7 +12,7 @@
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "pod-install": "pushd ios; bundle exec pod install; popd",
     "run-android": "react-native run-android --variant debug",
-    "run-ios": "react-native run-ios",
+    "run-ios": "react-native run-ios --scheme Staging",
     "start": "react-native start --reset-cache",
     "test": "jest --coverage",
     "test:ios": "detox test --configuration=ios",
@@ -56,6 +56,7 @@
     "react-native-background-fetch": "^3.1.0",
     "react-native-base64": "^0.2.1",
     "react-native-config": "^1.1.0",
+    "react-native-fs": "^2.16.6",
     "react-native-gesture-handler": "^1.6.1",
     "react-native-linear-gradient": "^2.5.6",
     "react-native-localize": "^1.4.0",
