+++ src/screens/qr/views/QRCodeScanner.tsx
@@ -5,7 +5,7 @@ import {Box, Text} from 'components';
 import {useI18n} from 'locale';
 import {useNavigation} from '@react-navigation/native';
 import {log} from 'shared/logging/config';
-import {useOutbreakService} from 'shared/OutbreakProvider';
+import {useOutbreakService} from 'services/OutbreakService';
 import {Icon} from 'components/Icon';
 
 import {Toolbar} from '../components/Toolbar';
