+++ src/services/OutbreakService/index.ts
@@ -1,0 +1,2 @@
+export * from './OutbreakProvider';
+export * from './OutbreakService';
